+++ test/options.cpp
@@ -8,7 +8,7 @@ class Argv {
   public:
 
   Argv(std::initializer_list<const char*> args)
-  : m_argv(new char*[args.size()])
+  : m_argv(new const char*[args.size()])
   , m_argc(static_cast<int>(args.size()))
   {
     int i = 0;
@@ -26,7 +26,7 @@ class Argv {
     }
   }
 
-  char** argv() const {
+  const char** argv() const {
     return m_argv.get();
   }
 
@@ -37,7 +37,7 @@ class Argv {
   private:
 
   std::vector<std::unique_ptr<char[]>> m_args;
-  std::unique_ptr<char*[]> m_argv;
+  std::unique_ptr<const char*[]> m_argv;
   int m_argc;
 };
 
@@ -69,7 +69,7 @@ TEST_CASE("Basic options", "[options]")
     "--space",
   });
 
-  char** actual_argv = argv.argv();
+  auto** actual_argv = argv.argv();
   auto argc = argv.argc();
 
   auto result = options.parse(argc, actual_argv);
@@ -94,7 +94,7 @@ TEST_CASE("Basic options", "[options]")
   CHECK(arguments[2].key() == "value");
   CHECK(arguments[3].key() == "av");
 
-  CHECK_THROWS_AS(result["nothing"].as<std::string>(), std::domain_error&);
+  CHECK_THROWS_AS(result["nothing"].as<std::string>(), cxxopts::option_has_no_value_exception&);
 }
 
 TEST_CASE("Short options", "[options]")
@@ -125,7 +125,7 @@ TEST_CASE("No positional", "[positional]")
 
   Argv av({"tester", "a", "b", "def"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
   auto result = options.parse(argc, argv);
 
@@ -154,7 +154,7 @@ TEST_CASE("All positional", "[positional]")
 
   auto result = options.parse(argc, argv);
 
-  REQUIRE(argc == 1);
+  CHECK(result.unmatched().size() == 0);
   REQUIRE(positional.size() == 3);
 
   CHECK(positional[0] == "a");
@@ -177,12 +177,12 @@ TEST_CASE("Some positional explicit", "[positional]")
 
   Argv av({"tester", "--output", "a", "b", "c", "d"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   auto result = options.parse(argc, argv);
 
-  CHECK(argc == 1);
+  CHECK(result.unmatched().size() == 0);
   CHECK(result.count("output"));
   CHECK(result["input"].as<std::string>() == "b");
   CHECK(result["output"].as<std::string>() == "a");
@@ -203,17 +203,16 @@ TEST_CASE("No positional with extras", "[positional]")
 
   Argv av({"extras", "--", "a", "b", "c", "d"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   auto old_argv = argv;
   auto old_argc = argc;
 
-  options.parse(argc, argv);
+  auto result = options.parse(argc, argv);
 
-  REQUIRE(argc == old_argc - 1);
-  CHECK(argv[0] == std::string("extras"));
-  CHECK(argv[1] == std::string("a"));
+  auto& unmatched = result.unmatched();
+  CHECK((unmatched == std::vector<std::string>{"a", "b", "c", "d"}));
 }
 
 TEST_CASE("Positional not valid", "[positional]") {
@@ -226,7 +225,7 @@ TEST_CASE("Positional not valid", "[positional]") {
 
   Argv av({"foobar", "bar", "baz"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   CHECK_THROWS_AS(options.parse(argc, argv), cxxopts::option_not_exists_exception&);
@@ -241,7 +240,7 @@ TEST_CASE("Empty with implicit value", "[implicit]")
 
   Argv av({"implicit", "--implicit="});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   auto result = options.parse(argc, argv);
@@ -260,7 +259,7 @@ TEST_CASE("Boolean without implicit value", "[implicit]")
   SECTION("When no value provided") {
     Argv av({"no_implicit", "--bool"});
 
-    char** argv = av.argv();
+    auto** argv = av.argv();
     auto argc = av.argc();
 
     CHECK_THROWS_AS(options.parse(argc, argv), cxxopts::missing_argument_exception&);
@@ -269,7 +268,7 @@ TEST_CASE("Boolean without implicit value", "[implicit]")
   SECTION("With equal-separated true") {
     Argv av({"no_implicit", "--bool=true"});
 
-    char** argv = av.argv();
+    auto** argv = av.argv();
     auto argc = av.argc();
 
     auto result = options.parse(argc, argv);
@@ -280,7 +279,7 @@ TEST_CASE("Boolean without implicit value", "[implicit]")
   SECTION("With equal-separated false") {
     Argv av({"no_implicit", "--bool=false"});
 
-    char** argv = av.argv();
+    auto** argv = av.argv();
     auto argc = av.argc();
 
     auto result = options.parse(argc, argv);
@@ -291,7 +290,7 @@ TEST_CASE("Boolean without implicit value", "[implicit]")
   SECTION("With space-separated true") {
     Argv av({"no_implicit", "--bool", "true"});
 
-    char** argv = av.argv();
+    auto** argv = av.argv();
     auto argc = av.argc();
 
     auto result = options.parse(argc, argv);
@@ -302,7 +301,7 @@ TEST_CASE("Boolean without implicit value", "[implicit]")
   SECTION("With space-separated false") {
     Argv av({"no_implicit", "--bool", "false"});
 
-    char** argv = av.argv();
+    auto** argv = av.argv();
     auto argc = av.argc();
 
     auto result = options.parse(argc, argv);
@@ -323,7 +322,7 @@ TEST_CASE("Default values", "[default]")
   SECTION("Sets defaults") {
     Argv av({"implicit"});
 
-    char** argv = av.argv();
+    auto** argv = av.argv();
     auto argc = av.argc();
 
     auto result = options.parse(argc, argv);
@@ -339,7 +338,7 @@ TEST_CASE("Default values", "[default]")
   SECTION("When values provided") {
     Argv av({"implicit", "--default", "5"});
 
-    char** argv = av.argv();
+    auto** argv = av.argv();
     auto argc = av.argc();
 
     auto result = options.parse(argc, argv);
@@ -374,7 +373,7 @@ TEST_CASE("Integers", "[options]")
 
   Argv av({"ints", "--", "5", "6", "-6", "0", "0xab", "0xAf", "0x0"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   options.parse_positional("positional");
@@ -401,7 +400,7 @@ TEST_CASE("Leading zero integers", "[options]")
 
   Argv av({"ints", "--", "05", "06", "0x0ab", "0x0001"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   options.parse_positional("positional");
@@ -425,7 +424,7 @@ TEST_CASE("Unsigned integers", "[options]")
 
   Argv av({"ints", "--", "-2"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   options.parse_positional("positional");
@@ -504,7 +503,7 @@ TEST_CASE("Floats", "[options]")
 
   Argv av({"floats", "--double", "0.5", "--", "4", "-4", "1.5e6", "-1.5e6"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   options.parse_positional("positional");
@@ -529,7 +528,7 @@ TEST_CASE("Invalid integers", "[integer]") {
 
     Argv av({"ints", "--", "Ae"});
 
-    char **argv = av.argv();
+    auto** argv = av.argv();
     auto argc = av.argc();
 
     options.parse_positional("positional");
@@ -554,7 +553,7 @@ TEST_CASE("Booleans", "[boolean]") {
 
   Argv av({"booleans", "--bool=false", "--debug=true", "--timing", "--verbose=1", "--dry-run=0", "extra"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   auto result = options.parse(argc, argv);
@@ -588,7 +587,7 @@ TEST_CASE("std::vector", "[vector]") {
 
   Argv av({"vector", "--vector", "1,-2.1,3,4.5"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   options.parse(argc, argv);
@@ -609,7 +608,7 @@ TEST_CASE("std::optional", "[optional]") {
 
   Argv av({"optional", "--optional", "foo"});
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   options.parse(argc, argv);
@@ -634,7 +633,7 @@ TEST_CASE("Unrecognised options", "[options]") {
     "--another_unknown",
   });
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   SECTION("Default behaviour") {
@@ -643,9 +642,9 @@ TEST_CASE("Unrecognised options", "[options]") {
 
   SECTION("After allowing unrecognised options") {
     options.allow_unrecognised_options();
-    CHECK_NOTHROW(options.parse(argc, argv));
-    REQUIRE(argc == 3);
-    CHECK_THAT(argv[1], Catch::Equals("--unknown"));
+    auto result = options.parse(argc, argv);
+    auto& unmatched = result.unmatched();
+    CHECK((unmatched == std::vector<std::string>{"--unknown", "--another_unknown"}));
   }
 }
 
@@ -661,7 +660,7 @@ TEST_CASE("Allow bad short syntax", "[options]") {
     "-some_bad_short",
   });
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   SECTION("Default behaviour") {
@@ -684,7 +683,7 @@ TEST_CASE("Invalid option syntax", "[options]") {
     "--a",
   });
 
-  char** argv = av.argv();
+  auto** argv = av.argv();
   auto argc = av.argc();
 
   SECTION("Default behaviour") {
@@ -704,7 +703,7 @@ TEST_CASE("Options empty", "[options]") {
        "--unknown"
      });
   auto argc = argv_.argc();
-  char** argv = argv_.argv();
+  auto** argv = argv_.argv();
 
   CHECK(options.groups().empty());
   CHECK_THROWS_AS(options.parse(argc, argv), cxxopts::option_not_exists_exception&);
@@ -733,7 +732,7 @@ TEST_CASE("Initializer list with group", "[options]") {
       "8000",
       "-t",
     });
-  char** actual_argv = argv.argv();
+  auto** actual_argv = argv.argv();
   auto argc = argv.argc();
   auto result = options.parse(argc, actual_argv);
 
@@ -763,7 +762,7 @@ TEST_CASE("Option add with add_option(string, Option)", "[options]") {
        "4"
      });
   auto argc = argv_.argc();
-  char** argv = argv_.argv();
+  auto** argv = argv_.argv();
   auto result = options.parse(argc, argv);
 
   CHECK(result.arguments().size()==2);